refactor(gateway): extract worker-placement and startup-trace modules from server.impl
Move createGatewayStartupTrace to server-startup-trace.ts and worker
placement bootstrap helpers to server-worker-placement.ts to keep
server.impl.ts under the LOC ratchet baseline.
